:recycle: Extract walkToGroup
